Export hueExtend for custom converters
