+++ lib/philips.js
@@ -2,6 +2,9 @@
 
 const ColorXY = require('./color').ColorXY;
 const ColorRGB = require('./color').ColorRGB;
+const extendDontUse = require('./extend');
+const ota = require('./ota');
+const exposes = require('./exposes');
 
 const encodeRGBToScaledGradient = (hex) => {
     const xy = ColorRGB.fromHex(hex).toXY();
@@ -41,6 +44,103 @@ const knownEffects = {
     '0a80': 'sparkle',
 };
 
+const hueExtend = {
+    light_onoff_brightness: (options={}) => {
+        options = {disableHueEffects: true, ...options};
+        if (!options.disableHueEffects) options.disableEffect = true;
+        const result = extendDontUse.light_onoff_brightness(options);
+        result['ota'] = ota.zigbeeOTA;
+        result['meta'] = {turnsOffAtBrightness1: true};
+        result['toZigbee'] = result['toZigbee'].concat([tzLocal.hue_power_on_behavior, tzLocal.hue_power_on_error]);
+        if (!options.disableHueEffects) {
+            result['toZigbee'] = result['toZigbee'].concat([tzLocal.effect]);
+            result['exposes'] = result['exposes'].concat([exposes.enum('effect', ea.SET,
+                ['blink', 'breathe', 'okay', 'channel_change', 'candle', 'finish_effect', 'stop_effect', 'stop_hue_effect'])]);
+        }
+        return result;
+    },
+    light_onoff_brightness_colortemp: (options={}) => {
+        options = {disableHueEffects: true, ...options};
+        if (!options.disableHueEffects) options.disableEffect = true;
+        const result = extendDontUse.light_onoff_brightness_colortemp(options);
+        result['ota'] = ota.zigbeeOTA;
+        result['meta'] = {turnsOffAtBrightness1: true};
+        result['toZigbee'] = result['toZigbee'].concat([tzLocal.hue_power_on_behavior, tzLocal.hue_power_on_error]);
+        if (!options.disableHueEffects) {
+            result['toZigbee'] = result['toZigbee'].concat([tzLocal.effect]);
+            result['exposes'] = result['exposes'].concat([exposes.enum('effect', ea.SET,
+                ['blink', 'breathe', 'okay', 'channel_change', 'candle', 'finish_effect', 'stop_effect', 'stop_hue_effect'])]);
+        }
+        return result;
+    },
+    light_onoff_brightness_color: (options={}) => {
+        options = {disableHueEffects: true, ...options};
+        if (!options.disableHueEffects) options.disableEffect = true;
+        const result = extendDontUse.light_onoff_brightness_color({supportsHS: true, ...options});
+        result['ota'] = ota.zigbeeOTA;
+        result['meta'] = {turnsOffAtBrightness1: true};
+        result['toZigbee'] = result['toZigbee'].concat([tzLocal.hue_power_on_behavior, tzLocal.hue_power_on_error]);
+        if (!options.disableHueEffects) {
+            result['toZigbee'] = result['toZigbee'].concat([tzLocal.effect]);
+            result['exposes'] = result['exposes'].concat([exposes.enum('effect', ea.SET, [
+                'blink', 'breathe', 'okay', 'channel_change',
+                'candle', 'fireplace', 'colorloop',
+                'finish_effect', 'stop_effect', 'stop_hue_effect',
+            ])]);
+        }
+        return result;
+    },
+    light_onoff_brightness_colortemp_color: (options={}) => {
+        options = {disableHueEffects: true, ...options};
+        if (!options.disableHueEffects) options.disableEffect = true;
+        const result = extendDontUse.light_onoff_brightness_colortemp_color({supportsHS: true, ...options});
+        result['ota'] = ota.zigbeeOTA;
+        result['meta'] = {turnsOffAtBrightness1: true};
+        result['toZigbee'] = result['toZigbee'].concat([tzLocal.hue_power_on_behavior, tzLocal.hue_power_on_error]);
+        if (!options.disableHueEffects) {
+            result['toZigbee'] = result['toZigbee'].concat([tzLocal.effect]);
+            result['exposes'] = result['exposes'].concat([exposes.enum('effect', ea.SET, [
+                'blink', 'breathe', 'okay', 'channel_change',
+                'candle', 'fireplace', 'colorloop',
+                'finish_effect', 'stop_effect', 'stop_hue_effect',
+            ])]);
+        }
+        return result;
+    },
+    light_onoff_brightness_colortemp_color_gradient: (options={}) => {
+        options = {supportsHS: true, disableEffect: true, extraEffects: [], ...options};
+        const result = extendDontUse.light_onoff_brightness_colortemp_color({noConfigure: true, ...options});
+        result['ota'] = ota.zigbeeOTA;
+        result['meta'] = {turnsOffAtBrightness1: true};
+        result['toZigbee'] = result['toZigbee'].concat([
+            tzLocal.hue_power_on_behavior, tzLocal.hue_power_on_error, tzLocal.effect,
+            tzLocal.gradient_scene, tzLocal.gradient({reverse: true}),
+        ]);
+        result['fromZigbee'] = result['fromZigbee'].concat([fzLocal.gradient({reverse: true})]);
+        result['configure'] = async (device, coordinatorEndpoint, logger) => {
+            await extendDontUse.light_onoff_brightness_colortemp_color(options)
+                .configure(device, coordinatorEndpoint, logger);
+            for (const ep of device.endpoints) {
+                await ep.bind('manuSpecificPhilips2', coordinatorEndpoint);
+            }
+        };
+        result['exposes'] = result['exposes'].concat([
+            // gradient_scene is deprecated, use gradient instead
+            exposes.enum('gradient_scene', ea.SET, Object.keys(gradientScenes)),
+            exposes.list('gradient', ea.ALL, exposes.text('hex', 'Color in RGB HEX format (eg #663399)'))
+                .withLengthMin(1)
+                .withLengthMax(9)
+                .withDescription('List of RGB HEX colors'),
+            exposes.enum('effect', ea.SET, [
+                'blink', 'breathe', 'okay', 'channel_change',
+                'candle', 'fireplace', 'colorloop', 'sunrise',
+                'finish_effect', 'stop_effect', 'stop_hue_effect',
+                ...options.extraEffects,
+            ]),
+        ]);
+        return result;
+    },
+};
 
 // decoder for manuSpecificPhilips2.state
 function decodeGradientColors(input, opts) {
@@ -249,4 +349,5 @@ function encodeGradientColors(value, opts) {
 module.exports = {
     decodeGradientColors,
     encodeGradientColors,
+    hueExtend,
 };
